Moved exporter, removed limit, refactored component
